+++ autoprotocol/container.py
@@ -394,15 +394,14 @@ class Well(EntityPropertiesMixin):
         ------
         TypeError
             Incorrect input-type given
-        RuntimeError
-            Compound information does not contain required keys
         """
-        expected_keys = {
-            "id",
-            "molecularWeight",
-            "smiles",
-            "concentration",
-            "solubilityFlag",
+        # expected parameters and label transformations
+        expected_params = {
+            "id": "id",
+            "molecularWeight": "molecular_weight",
+            "smiles": "smiles",
+            "concentration": "concentration",
+            "solubilityFlag": "solubility_flag",
         }
         if not isinstance(compounds, list):
             raise TypeError(
@@ -410,19 +409,32 @@ class Well(EntityPropertiesMixin):
             )
 
         for compound in compounds:
-            # Check all expected keys aer present
-            if set(compound.keys()) != expected_keys:
-                raise RuntimeError(
-                    "Compound information must include `id`, `molecularWeight`, `concentration`, `solubilityFlag` and `smiles` keys."
-                )
-            # Transform {"molecularWeight": float} -> {"molecular_weight": Unit}
-            compound["molecular_weight"] = Unit(
-                compound.pop("molecularWeight"), "g/mol"
-            )
-            compound["solubility_flag"] = compound.pop("solubilityFlag")
-            compound["concentration"] = Unit(
-                compound.pop("concentration"), "millimol/liter"
-            )
+            if isinstance(compound, dict):
+                for k in expected_params:
+                    if k in compound:
+                        # transform {"molecularWeight": float} -> {"molecular_weight": Unit}
+                        if k == "molecularWeight":
+                            compound[expected_params[k]] = Unit(
+                                compound.pop("molecularWeight"), "g/mol"
+                            )
+                        elif k == "solubilityFlag":
+                            compound[expected_params[k]] = compound.pop(
+                                "solubilityFlag"
+                            )
+                        elif k == "concentration":
+                            compound[expected_params[k]] = Unit(
+                                compound.pop("concentration"), "millimol/liter"
+                            )
+                        elif k == "smiles":
+                            # TODO: validation on smiles
+                            pass
+                        else:
+                            pass
+                    else:
+                        # set unspecified keys using preferred param string
+                        compound[expected_params[k]] = None
+            else:
+                pass
 
         self.compounds = compounds
         return self